Add files via upload
This is a Basic UI that is developed by me. The UI explores and Eloborates the Election Page of India  where I made the Basic Navigation on 
Uppermost  Side and Made a Small User Registration before submitting the age  Such that User Can Fill All the Required Information Before Vote. 
I mentioned A notice that if Age of User less than to 18 They Have to wait 18 and I Gave The crucial Information for Their Voter Id Formation Of the 
Age of The User is > 18 Then They Can easily Vote.